# bin/docs/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: Doppler App REST API
host: localhost:8080
basePath: /api/v1
schemes:
  - http
  - https
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /events:
    get:
      summary: search events
      tags:
        - Event
      security:
        - Bearer: []
      parameters:
        - in: query
          name: limit
          type: integer
          description: the search limit. Must be greater than 0 if present
        - in: query
          name: offset
          type: integer
          description: the search offset. Must be not less than 0 if present
        - in: query
          name: sortBy
          type: string
          description: the sort field. Allow `firstSessionStart` (sort by the first session's start date)
        - in: query
          name: sortDirection
          type: string
          description: the sort direction. Allow `asc` and `desc`. Default to `asc` if not present
        - in: query
          name: topicId
          type: string
          description: the topic UUID filter
        - in: query
          name: status
          type: string
          enum:
            - PAST
            - UPCOMING
          description: the status filter. If `PAST`, filter only the completed events. If `UPCOMING`, filter only the events which have not been completed
        - in: query
          name: onlyMyEvents
          type: boolean
          description: the onlyMyEvents filter. True to filter only the events to which the current users registered. Default to false if not present
      responses:
        200:
          description: successful operation
          schema:
            type: object
            properties:
              total:
                type: integer
              items:
                type: array
                items:
                  $ref: "#/definitions/Event"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
  /events/{eventId}:
    get:
      summary: get an event
      tags:
        - Event
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: the event UUID
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/Event"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
  /events/{eventId}/share:
    post:
      summary: share the event to other users
      description: >
         - share the event by creating new notifications for the specified users about the event
         - the notification message should be configurable
      tags:
        - Event
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: the event UUID
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              properties:
                userId:
                  type: string
      responses:
        204:
          description: successful operation
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
...
  /events/{eventId}/comments:
    parameters:
    - in: path
      name: eventId
      required: true
      type: string
      description: the event UUID

    post:
      summary: create a new comment
      tags:
        - Event Comment
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/EventCommentRequest"
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/EventComment"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
    get:
      summary: search event comments
      tags:
        - Event Comment
      security:
        - Bearer: []
      parameters:
        - in: query
          name: limit
          type: integer
          description: the search limit. Must be greater than 0 if present
        - in: query
          name: offset
          type: integer
          description: the search offset. Must be not less than 0 if present
        - in: query
          name: sortBy
          type: string
          description: the sort field. Allow `createdAt` and `likeCount`
        - in: query
          name: sortDirection
          type: string
          description: the sort direction. Allow `asc` and `desc`. Default to `asc` if not present
      responses:
        200:
          description: successful operation
          schema:
            type: object
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: "#/definitions/EventComment"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
  /events/{eventId}/comments/{commentId}:
    parameters:
    - in: path
      name: eventId
      required: true
      type: string
      description: the event UUID
    - in: path
      name: commentId
      required: true
      type: string
      description: the event comment UUID
    
    put:
      summary: update an event comment
      tags:
        - Event Comment
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/EventCommentRequest"
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/EventComment"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        403:
          description: the user is not allowed to update another user's comment | not allowed to update deleted comment
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event or event comment does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"

    delete:
      summary: delete an event comment (soft delete)
      description: >
        - soft delete using a delete flag 
        
        - set the comment message to "This comment has been deleted" so that no one can recover the message
      tags:
        - Event Comment
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/EventComment"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        403:
          description: the user is not allowed to delete another user's comment | not allowed to delete deleted comment
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event or event comment does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
  /events/{eventId}/comments/{commentId}/like:
    parameters:
    - in: path
      name: eventId
      required: true
      type: string
      description: the event UUID
    - in: path
      name: commentId
      required: true
      type: string
      description: the event comment UUID

    post:
      summary: like an event comment (by the logged-in user)
      tags:
        - Event Comment
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/EventComment"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        403:
          description: the user is not allowed to like his/her own comment | not allowed to like deleted comment
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event or event comment does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
  /events/{eventId}/comments/{commentId}/unlike:
    parameters:
    - in: path
      name: eventId
      required: true
      type: string
      description: the event UUID
    - in: path
      name: commentId
      required: true
      type: string
      description: the event comment UUID

    post:
      summary: unlike an event comment (by the logged-in user)
      tags:
        - Event Comment
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/EventComment"
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        403:
          description: the user is not allowed to unlike deleted comment
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event or event comment does not exist | the user has not liked the event comment yet
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
...
  /events/{eventId}/quiz:
    get:
      summary: get the quiz questions of the event for the logged-in user
      tags:
        - Quiz
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          type: string
          required: true
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/QuizQuestion'
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        403:
          description: the user is not allowed to get quiz questions of the event he/she has not attended
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
            
    post:
      summary: submit the quiz answers of the event for the logged-in user
      description: >
        - Add a new `UserRewardPoint` for completing the quiz, the points should be got from the event or configuration
      
        - Add the points to the `User.points`
        
        - Return the `UserRewardPoint`
      tags:
        - Quiz
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/QuizAnswer'
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/UserRewardPoint'
        400:
          description: invalid request
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          description: action is not allowed for anonymous | invalid or expired access token
          schema:
            $ref: "#/definitions/ErrorMessage"
        403:
          description: the user is not allowed to submit quiz answers for the event he/she has not attended
          schema:
            $ref: "#/definitions/ErrorMessage"
        404:
          description: the event does not exist
          schema:
            $ref: "#/definitions/ErrorMessage"
        500:
          description: server error
          schema:
            $ref: "#/definitions/ErrorMessage"
...
definitions:
  ErrorMessage:
    type: object
    properties:
      message:
        type: string
...
  User:
    type: object
    properties:
      id:
        type: string
        description: the UUID
      email:
        type: string
      fullName:
        type: string
        description: synced from SAML
      division:
        type: string
        description: synced from SAML
      position:
        type: string
        description: synced from SAML
      location:
        type: string
        description: synced from SAML
      phoneNumber:
        type: string
        description: synced from SAML
      photoUrl:
        type: string
        description: synced from SAML
      isNew:
        type: boolean
        description: the flag to indicate that the user is a new or an existing user, if the user is a new user, client side should display the sign-up screens so that the user can select the topic references
      points:
        type: integer
        minimum: 0
        description: the current points
...
  UserRewardPoint:
    type: object
    description: the history of points earned by the user
    properties:
      id:
        type: string
        description: the UUID, should not be present in the request body
      description:
        type: string
      points:
        type: number
        format: integer
        minimum: 1
      createdAt:
        type: string
        format: date
        description: the created date time, generated by the system and should not be present in the request body
      userId:
        type: string
        description: the user UUID
    required:
      - description
      - points
      - createdAt
      - userId
  Event:
    type: object
    properties:
      id:
        type: string
        description: the UUID, should not be present in the request body
      title:
        type: string
      description:
        type: string
      pointsForRegistering:
        type: number
        format: integer
        minimum: 1
        description: the points that each attendee earns for registering to the event
      pointsForScanningTicket:
        type: number
        format: integer
        minimum: 1
        description: the points that each attendee earns for scanning ticket
      topicId:
        type: string
      location:
        type: string
      currentBooking:
        type: integer
        description: the current booking
      maxCapacity:
        type: integer
        description: the max capacity
      waitingCapacity:
        type: integer
        description: the waiting capacity
      tags:
        type: array
        items:
          type: string
      sessions:
        type: array
        items:
          $ref: '#/definitions/EventSession'
      portraitImageUrl:
        type: string
        description: the portrait image url (to be displayed in the event list)
      landscapeImageUrl:
        type: string
        description: the landscape image url (to be displayed in the event detail page)
      isMyEvent:
        type: boolean
        description: the flag to indicate whether the current logged-in user has registered to the event
  EventSession:
    type: object
    properties:
      id:
        type: string
        description: the UUID, should not be present in the request body
      title:
        type: string
      location:
        type: string
      start:
        type: string
        format: date
      end:
        type: string
        format: date
      speaker:
        type: object
        properties:
          name:
            type: string
          photoUrl:
            type: string
  EventComment:
    type: object
    properties:
      id:
        type: string
        description: the event comment UUID, should not be present in the request body
      content:
        type: string
      user:
        description: the user to create the comment, should not be present in the request body
        $ref: '#/definitions/User'
      createdAt:
        type: string
        description: the created date time, generated by the system and should not be present in the request body
      likeCount:
        type: integer
        description: the number of likes
      likedByMe:
        type: boolean
        description: the flag to indicate that the comment is liked by the current logged-in user or not
      childComments:
        type: array
        items:
          $ref: '#/definitions/EventComment'
        description: the child comments, not that this app supports only 2 levels of comments (root level and child level) 
      deleted:
        type: boolean
        description: the delete flag
    required:
      - content
      - user
      - createdAt
      - likeCount
      - likedByMe
  EventCommentRequest:
    type: object
    properties:
      content:
        type: string
      parentCommentId:
        type: string
        description: the parent comment UUID, null if this is root comment, supporting only 2 levels of comments
    required:
      - content
...
  QuizQuestion:
    type: object
    properties:
      id:
        type: string
      question:
        type: string
      type:
        type: string
        enum:
          - text
          - singleChoice
          - multipleChoice
      quizAnswerOptions:
        type: array
        items:
          $ref: '#/definitions/QuizAnswerOption'
        description: 0 items if the type is `text`
    required:
      - question
      - type
  QuizAnswerOption:
    type: object
    properties:
      id:
        type: string
      answerOption:
        type: string
    required:
      - quizQuestionId
      - answerOption
  QuizAnswer:
    type: object
    properties:
      id:
        type: string
        description: the quiz answer UUID, should not be present in the request body
      quizQuestionId:
        type: string
      selectedQuizAnswerOptionId:
        type: string
        description: null if the question type is `text`
      textAnswer:
        type: string
        description: not null if the question type is `text`
    required:
      - quizQuestionId
...